style(msvc): fix C4244/C4267 in XmlToLua/WheelBase/StatsManager/ScreenOptions/Course/ActorMultiTexture (item 2)
All static_cast, documenting existing intentional narrowing -- no
logic changes. sm_tests unchanged at 5966/226.

- XmlToLua.cpp (6): RString::Left/Right fed size_t arithmetic at 5
  call sites parsing field names and file paths
  (add_extension_to_relative_path..., store_cmd, store_field,
  load_frames_from_file x2); a trailing-.lua filename Left() call.
- WheelBase.cpp (6): a size_t item count into int; five
  wrap(int&, size_t) sites across ChangeMusic/RebuildWheelItems/
  FirstVisibleIndex.
- StatsManager.cpp (6): a size_t song count into unsigned; a
  std::trunc() result and two size_t accumulators feeding profile
  gameplay-seconds/songs-played counters; a size_t-minus-int index; a
  lua_pushnumber(size_t) accessor.
- ScreenOptions.cpp (6): a MoveRowAbsolute call's size_t-minus-1 arg;
  the same 'const int iNumChoices = ...m_vsChoices.size()' pattern at
  2 call sites (4 warnings, both int/const int variants); a
  wrap(int&, size_t) site.
- Course.cpp (6): a size_t mod-change count into int; two
  RandomInt(size_t) calls choosing steps; two identical
  std::floor((iMinDist + iMaxDist) / 2) sites promoting an all-int
  expression to double then narrowing back, in two near-duplicate
  meter-balancing code paths; a lua_pushnumber(size_t) accessor.
- ActorMultiTexture.cpp (6): two int-into-float texture-size members;
  two size_t-into-int texture-unit-count returns; two
  enum_add2(TextureUnit_1, size_t) calls needing an int second arg.

+++ src/ActorMultiTexture.cpp
@@ -52,8 +52,8 @@ void ActorMultiTexture::LoadFromNode( const XNode* pNode )
 
 void ActorMultiTexture::SetSizeFromTexture( RageTexture *pTexture )
 {
-	ActorMultiTexture::m_size.x = pTexture->GetSourceWidth();
-	ActorMultiTexture::m_size.y = pTexture->GetSourceHeight();
+	ActorMultiTexture::m_size.x = static_cast<float>(pTexture->GetSourceWidth());
+	ActorMultiTexture::m_size.y = static_cast<float>(pTexture->GetSourceHeight());
 }
 
 void ActorMultiTexture::ClearTextures()
@@ -68,13 +68,13 @@ int ActorMultiTexture::AddTexture( RageTexture *pTexture )
 	if( pTexture == nullptr )
 	{
 		LOG->Warn( "Can't add nil texture to ActorMultiTexture" );
-		return m_aTextureUnits.size();
+		return static_cast<int>(m_aTextureUnits.size());
 	}
 	LOG->Trace( "ActorMultiTexture::AddTexture( %s )", pTexture->GetID().filename.c_str() );
 
 	m_aTextureUnits.push_back( TextureUnitState() );
 	m_aTextureUnits.back().m_pTexture = TEXTUREMAN->CopyTexture( pTexture );
-	return m_aTextureUnits.size();
+	return static_cast<int>(m_aTextureUnits.size());
 }
 
 void ActorMultiTexture::SetTextureMode( int iIndex, TextureMode tm )
@@ -100,7 +100,7 @@ void ActorMultiTexture::DrawPrimitives()
 	DISPLAY->ClearAllTextures();
 	for( std::size_t i = 0; i < m_aTextureUnits.size(); ++i )
 	{
-		TextureUnit tu = enum_add2(TextureUnit_1, i);
+		TextureUnit tu = enum_add2(TextureUnit_1, static_cast<int>(i));
 		DISPLAY->SetTexture( tu, m_aTextureUnits[i].m_pTexture->GetTexHandle() );
 		DISPLAY->SetTextureWrapping( tu, m_bTextureWrapping );
 		DISPLAY->SetTextureMode( tu, m_aTextureUnits[i].m_TextureMode );
@@ -128,7 +128,7 @@ void ActorMultiTexture::DrawPrimitives()
 	DISPLAY->DrawQuad( v );
 
 	for( std::size_t i = 0; i < m_aTextureUnits.size(); ++i )
-		DISPLAY->SetTexture( enum_add2(TextureUnit_1, i), 0 );
+		DISPLAY->SetTexture( enum_add2(TextureUnit_1, static_cast<int>(i)), 0 );
 
 	DISPLAY->SetEffectMode( EffectMode_Normal );
 }
